requirements.txt
altair==4.1.0
altair-saver==0.5.0
APScheduler==3.6.3
feedparser==6.0.2
pandas==1.1.5
PyYAML==5.3.1
requests==2.25.1
sqlitedict==1.7.0
tweepy==3.9.0